Consolidate entrypoint
