+++ Dockerfile.optimized
@@ -114,7 +114,7 @@ HEALTHCHECK --interval=30s --timeout=10s --start-period=40s --retries=3 \
     CMD curl -f http://localhost:8000/api/v1/system/health || exit 1
 
 # Default command with optimized settings
-CMD ["python", "-m", "uvicorn", "main_v2:app", "--host", "0.0.0.0", "--port", "8000", "--workers", "1", "--log-level", "info"]
+CMD ["python", "-m", "uvicorn", "main:app", "--host", "0.0.0.0", "--port", "8000", "--workers", "1", "--log-level", "info"]
 
 # Stage 4: Development runtime (optional)
 FROM production as development
@@ -140,4 +140,4 @@ RUN pip install --no-cache-dir \
 USER flightio
 
 # Development command with auto-reload
-CMD ["python", "-m", "uvicorn", "main_v2:app", "--host", "0.0.0.0", "--port", "8000", "--reload", "--log-level", "debug"] 
+CMD ["python", "-m", "uvicorn", "main:app", "--host", "0.0.0.0", "--port", "8000", "--reload", "--log-level", "debug"]
