+++ Makefile.optimized
@@ -261,12 +261,12 @@ docker-cleanup:
 # Start development server
 dev-server:
 	@echo -e "${BLUE}Starting development server...${NC}"
-	$(PYTHON) -m uvicorn main_v2:app --reload --host 0.0.0.0 --port 8000
+	$(PYTHON) -m uvicorn main:app --reload --host 0.0.0.0 --port 8000
 
 # Start production server
 prod-server:
 	@echo -e "${BLUE}Starting production server...${NC}"
-	$(PYTHON) -m uvicorn main_v2:app --host 0.0.0.0 --port 8000 --workers 4
+	$(PYTHON) -m uvicorn main:app --host 0.0.0.0 --port 8000 --workers 4
 
 # Start frontend development server
 dev-frontend:
